+++ src/wai/annotations/core/__init__.py
@@ -10,5 +10,6 @@ from ._InlineStreamProcessor import InlineStreamProcessor
 from ._InternalFormatConverter import InternalFormatConverter
 from ._Reader import Reader
 from ._SeparateImageWriter import SeparateImageWriter
+from ._Settings import Settings, get_settings, set_settings
 from ._typing import InternalFormat
 from ._Writer import Writer