+++ src/wai/annotations/core/utils/_get_associated_image.py
@@ -1,8 +1,6 @@
 import os
 from typing import Optional
 
-from .._ImageFormat import ImageFormat
-
 
 def get_associated_image(filename: str) -> Optional[str]:
     """
@@ -13,7 +11,8 @@ def get_associated_image(filename: str) -> Optional[str]:
     :return:            The filename of the found image, None if not found.
     """
     # Check each variation of the extension for each format
-    for image_format in ImageFormat:
+    from .._Settings import get_settings
+    for image_format in get_settings().IMAGE_FORMAT_PREFERENCE_ORDER:
         for extension in image_format.value:
             # Get the hypothetical filename for an image of this format
             image_filename: str = f"{filename}.{extension}"